Atualizados para etapa 4

+++ src/classesDasEntidades/atividades/Atividade.java
@@ -9,7 +9,7 @@ public abstract class Atividade implements Serializable{
 	protected DateFormat df = new SimpleDateFormat ("dd/MM/yyyy HH:mm");
 	protected String nome;
     protected boolean sinc;
-    public Map<Long, Integer> avaliacao = new HashMap<>(); //<matricula do estudante, avaliacao dele>
+    public Map<Long, Double> avaliacao = new HashMap<>(); //<matricula do estudante, avaliacao dele>
  
     public String getNome() {
 		return nome;
